Fix the version stamp set during build
The ldflags passed to the build expect to be the root package for the
repository, and should contain the pkg directory with the version
variables so that the build can set them.

The package passed was to the plugin entry point (./cmd/kubernetes)
which caused the version stamp to be empty.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>